refactor(plugin-trello&argocdapp): to adopt new plugin scaffold
