+++ internal/pkg/plugin/argocdapp/validate.go
@@ -2,6 +2,7 @@ package argocdapp
 
 import (
 	"fmt"
+
 	"k8s.io/apimachinery/pkg/util/validation"
 )
