+++ internal/pkg/plugin/trello/create.go
@@ -11,10 +11,12 @@ import (
 // Create creates Tello board and lists(todo/doing/done).
 func Create(options map[string]interface{}) (map[string]interface{}, error) {
 	var opts Options
+
 	if err := mapstructure.Decode(options, &opts); err != nil {
 		return nil, err
 	}
 
+	// validate parameters
 	if errs := validate(&opts); len(errs) != 0 {
 		for _, e := range errs {
 			log.Errorf("Options error: %s.", e)
